POWR-3735 Reconcile social media fields between group types
Unblocks POWR-3507.

These fields are not yet displayed, but will allow for the migration of the data from contact social media fields to their respective group social media fields, in order for the fields to be eventually removed from Contact once the design work is done for the group fields.

- Add LinkedIn/Nextdoor/YouTube to Elected official and Program
- Add all social media fields to Project and Advisory group (previously had none)
- Make order of contact and social media fields consistent between group types