PGOV-1019 update diff settings for document fields
Looks like this file hasn't been updated in a while so it's a huge change with all the new fields since then.

However, the only thing actually changed here is enabling the "Show file description" checkbox on all the file fields. This allows some actual context when looking at a revision diff where one of the files is deleted.