PGOV-1203 remove portland_media_embed_html_filter and install ckeditor5_pastefilter plugin (#3192)
* PGOV-1203 Update text filters to be more specific and only parse HTML when needed

* PGOV-1203 remove portland_media_embed_html_filter

This filter removes empty (or whitespace-only) elements in order to cleanup the HTML which is an admirable goal, but it causes confusing differences between the editor and the rendered page.

It can also cause serious performance issues on pages with complex nested HTML or many span tags (like those pasted from Word).

* PGOV-1203 install ckeditor5_pastefilter plugin and configure rules